moved user settings from /temp to /user
