+++ breadbox/breadbox/api/temp/__init__.py
@@ -2,7 +2,6 @@ from .router import router
 
 # import the following which register endpoints onto `router` as a side effect of being imported
 import breadbox.api.temp.cas
-import breadbox.api.temp.user_settings
 import breadbox.api.temp.context
 import breadbox.api.temp.associations
